Revisit XVRL reports
Fix #257

The p:validate-with-schematron step now returns a single xvrl:report. The bug
where digests weren’t “rolling up” correctly has been fixed.
