+++ xmlcalabash/src/main/kotlin/com/xmlcalabash/xvrl/XvrlReport.kt
@@ -1,13 +1,23 @@
 package com.xmlcalabash.xvrl
 
 import com.xmlcalabash.namespace.NsSaxon
+import com.xmlcalabash.namespace.NsSvrl
+import com.xmlcalabash.namespace.NsXml
 import com.xmlcalabash.namespace.NsXvrl
 import com.xmlcalabash.runtime.XProcStepConfiguration
 import com.xmlcalabash.util.SaxonTreeBuilder
 import com.xmlcalabash.util.Verbosity
+import com.xmlcalabash.xvrl.XvrlReports.Companion._context
+import com.xmlcalabash.xvrl.XvrlReports.Companion._document
+import com.xmlcalabash.xvrl.XvrlReports.Companion._documents
+import com.xmlcalabash.xvrl.XvrlReports.Companion._location
+import com.xmlcalabash.xvrl.XvrlReports.Companion._role
+import net.sf.saxon.om.NamespaceUri
+import net.sf.saxon.s9api.Axis
 import net.sf.saxon.s9api.HostLanguage
 import net.sf.saxon.s9api.QName
 import net.sf.saxon.s9api.XdmNode
+import net.sf.saxon.s9api.XdmNodeKind
 import net.sf.saxon.s9api.XmlProcessingError
 import net.sf.saxon.trans.XPathException
 import org.apache.logging.log4j.kotlin.logger
@@ -25,6 +35,12 @@ class XvrlReport private constructor(stepConfig: XProcStepConfiguration, val met
         fun newInstance(stepConfig: XProcStepConfiguration, metadata: XvrlReportMetadata, attr: Map<QName,String?> = emptyMap()): XvrlReport {
             val report = XvrlReport(stepConfig, metadata)
             report.commonAttributes(attr)
+            return report
+        }
+
+        fun fromSvrl(stepConfig: XProcStepConfiguration, svrl: XdmNode): XvrlReport {
+            val report = newInstance(stepConfig)
+            report.fromSvrl(svrl)
             return report
         }
     }
@@ -97,9 +113,9 @@ class XvrlReport private constructor(stepConfig: XProcStepConfiguration, val met
     }
 
     override fun serialize(builder: SaxonTreeBuilder) {
-        val digest = XvrlDigest(stepConfig)
         builder.addStartElement(NsXvrl.report, stepConfig.attributeMap(attributes))
         metadata.serialize(builder)
+        digest.clear()
         var overall = "true"
         for (item in detections) {
             item.serialize(builder)
@@ -138,4 +154,158 @@ class XvrlReport private constructor(stepConfig: XProcStepConfiguration, val met
         digest.serialize(builder)
         builder.addEndElement()
     }
+
+    // ============================================================
+
+    private fun fromSvrl(svrl: XdmNode) {
+        when (svrl.nodeKind) {
+            XdmNodeKind.DOCUMENT -> {
+                for (node in svrl.axisIterator(Axis.CHILD)) {
+                    fromSvrl(node)
+                }
+            }
+
+            XdmNodeKind.ELEMENT -> {
+                if (svrl.nodeName == NsSvrl.schematronOutput) {
+                    reportFromSvrl(this, svrl)
+                } else {
+                    for (node in svrl.axisIterator(Axis.CHILD)) {
+                        fromSvrl(node)
+                    }
+                }
+            }
+            else -> Unit
+        }
+    }
+
+    private fun reportFromSvrl(report: XvrlReport, svrl: XdmNode) {
+        report.metadata.schema(null, NamespaceUri.of("http://purl.oclc.org/dsdl/schematron"))
+        metadataFromSvrl(report.metadata, svrl)
+        val children = mutableListOf<XdmNode>()
+        for (node in svrl.axisIterator(Axis.CHILD)) {
+            if (node.nodeKind == XdmNodeKind.ELEMENT) {
+                children.add(node)
+            }
+        }
+
+        while (children.isNotEmpty()) {
+            val node = children.removeFirst()
+            if (node.nodeName == NsSvrl.activePattern) {
+                patternFromSvrl(report, node, children)
+            }
+        }
+    }
+
+    private fun metadataFromSvrl(metadata: XvrlReportMetadata, svrl: XdmNode) {
+        val seen = mutableSetOf<URI>()
+        for (node in svrl.axisIterator(Axis.CHILD)) {
+            when (node.nodeName) {
+                NsSvrl.activePattern -> {
+                    val documents = node.getAttributeValue(_documents) ?: node.getAttributeValue(_document)
+                    if (documents != null) {
+                        val uri = URI(documents)
+                        if (uri !in seen) {
+                            metadata.document(uri)
+                            seen.add(uri)
+                        }
+                    }
+                }
+                else -> Unit
+            }
+        }
+    }
+
+    private fun patternFromSvrl(report: XvrlReport, pattern: XdmNode, children: MutableList<XdmNode>) {
+        while (children.isNotEmpty()) {
+            if (children.first().nodeName == NsSvrl.activePattern) {
+                return
+            }
+            val node = children.removeFirst()
+            if (node.nodeName == NsSvrl.firedRule) {
+                ruleFromSvrl(report, pattern, node, children)
+            }
+        }
+    }
+
+    private fun ruleFromSvrl(report: XvrlReport, pattern: XdmNode, rule: XdmNode, children: MutableList<XdmNode>) {
+        while (children.isNotEmpty()) {
+            if (children.first().nodeName == NsSvrl.firedRule) {
+                return
+            }
+            val node = children.removeFirst()
+            when (node.nodeName) {
+                NsSvrl.successfulReport -> {
+                    detectionFromSvrl(report, pattern, rule, node, "info")
+                }
+                NsSvrl.failedAssert -> {
+                    detectionFromSvrl(report, pattern, rule, node, "error")
+                }
+                else -> Unit
+            }
+        }
+    }
+
+    private fun detectionFromSvrl(report: XvrlReport, pattern: XdmNode, rule: XdmNode, node: XdmNode, defaultSeverity: String) {
+        val severity = node.getAttributeValue(_role) ?: defaultSeverity
+        val detection = report.detection(severity)
+
+        if (node.getAttributeValue(_location) != null) {
+            val location = detection.location()
+            location.xpath = node.getAttributeValue(_location)
+        }
+
+        if (rule.getAttributeValue(_context) != null) {
+            val context = detection.context()
+            val location = context.location()
+            location.xpath = rule.getAttributeValue(_context)
+        }
+
+        val lang = node.getAttributeValue(NsXml.lang)
+        for (child in node.axisIterator(Axis.CHILD)) {
+            if (child.nodeName == NsSvrl.text || child.nodeName == NsSvrl.diagnosticReference) {
+                messageFromSvrl(detection, child, lang)
+            }
+        }
+    }
+
+    private fun messageFromSvrl(detection: XvrlDetection, node: XdmNode, defaultLanguage: String?) {
+        val lang = node.getAttributeValue(NsXml.lang) ?: defaultLanguage
+        val atts = mutableMapOf<QName,String>()
+        lang?.let { atts[NsXml.lang] = it }
+        val message = detection.message(atts)
+        for (child in node.axisIterator(Axis.CHILD)) {
+            when (child.nodeKind) {
+                XdmNodeKind.TEXT -> {
+                    val text = child.stringValue.trim()
+                    if (text.isNotEmpty()) {
+                        message.message(text)
+                    }
+                }
+                XdmNodeKind.ELEMENT -> {
+                    val element = message.message(child.nodeName)
+                    recurse(element, child)
+                }
+                else -> Unit
+            }
+        }
+    }
+
+    private fun recurse(message: XvrlMessageElement, node: XdmNode) {
+        for (child in node.axisIterator(Axis.CHILD)) {
+            when (child.nodeKind) {
+                XdmNodeKind.TEXT -> {
+                    val text = child.stringValue.trim()
+                    if (text.isNotEmpty()) {
+                        message.addContent(XvrlText(stepConfig, text))
+                    }
+                }
+                XdmNodeKind.ELEMENT -> {
+                    val element = message.message(child.nodeName)
+                    recurse(element, child)
+                }
+                else -> Unit
+            }
+        }
+    }
+
 }
