+++ xmlcalabash/src/main/kotlin/com/xmlcalabash/xvrl/XvrlReports.kt
@@ -40,12 +40,6 @@ class XvrlReports private constructor(stepConfig: XProcStepConfiguration): XvrlE
             reports._metadata = metadata
             return reports
         }
-
-        fun fromSvrl(stepConfig: XProcStepConfiguration, svrl: XdmNode): XvrlReports {
-            val reports = newInstance(stepConfig)
-            reports.fromSvrl(svrl)
-            return reports
-        }
     }
 
     fun report(attr: Map<QName,String?> = emptyMap()): XvrlReport {
@@ -159,159 +153,5 @@ class XvrlReports private constructor(stepConfig: XProcStepConfiguration): XvrlE
         }
         digest.serialize(builder)
         builder.addEndElement()
-    }
-
-    // ============================================================
-
-    private fun fromSvrl(svrl: XdmNode) {
-        when (svrl.nodeKind) {
-            XdmNodeKind.DOCUMENT -> {
-                for (node in svrl.axisIterator(Axis.CHILD)) {
-                    fromSvrl(node)
-                }
-            }
-
-            XdmNodeKind.ELEMENT -> {
-                if (svrl.nodeName == NsSvrl.schematronOutput) {
-                    reportFromSvrl(this, svrl)
-                } else {
-                    for (node in svrl.axisIterator(Axis.CHILD)) {
-                        fromSvrl(node)
-                    }
-                }
-            }
-            else -> Unit
-        }
-    }
-
-    private fun reportFromSvrl(reports: XvrlReports, svrl: XdmNode) {
-        val report = reports.report()
-        report.metadata.schema(null, NamespaceUri.of("http://purl.oclc.org/dsdl/schematron"))
-        metadataFromSvrl(report.metadata, svrl)
-        val children = mutableListOf<XdmNode>()
-        for (node in svrl.axisIterator(Axis.CHILD)) {
-            if (node.nodeKind == XdmNodeKind.ELEMENT) {
-                children.add(node)
-            }
-        }
-
-        while (children.isNotEmpty()) {
-            val node = children.removeFirst()
-            if (node.nodeName == NsSvrl.activePattern) {
-                patternFromSvrl(report, node, children)
-            }
-        }
-    }
-
-    private fun metadataFromSvrl(metadata: XvrlReportMetadata, svrl: XdmNode) {
-        val seen = mutableSetOf<URI>()
-        for (node in svrl.axisIterator(Axis.CHILD)) {
-            when (node.nodeName) {
-                NsSvrl.activePattern -> {
-                    val documents = node.getAttributeValue(_documents) ?: node.getAttributeValue(_document)
-                    if (documents != null) {
-                        val uri = URI(documents)
-                        if (uri !in seen) {
-                            metadata.document(uri)
-                            seen.add(uri)
-                        }
-                    }
-                }
-                else -> Unit
-            }
-        }
-    }
-
-    private fun patternFromSvrl(report: XvrlReport, pattern: XdmNode, children: MutableList<XdmNode>) {
-        while (children.isNotEmpty()) {
-            if (children.first().nodeName == NsSvrl.activePattern) {
-                return
-            }
-            val node = children.removeFirst()
-            if (node.nodeName == NsSvrl.firedRule) {
-                ruleFromSvrl(report, pattern, node, children)
-            }
-        }
-    }
-
-    private fun ruleFromSvrl(report: XvrlReport, pattern: XdmNode, rule: XdmNode, children: MutableList<XdmNode>) {
-        while (children.isNotEmpty()) {
-            if (children.first().nodeName == NsSvrl.firedRule) {
-                return
-            }
-            val node = children.removeFirst()
-            when (node.nodeName) {
-                NsSvrl.successfulReport -> {
-                    detectionFromSvrl(report, pattern, rule, node, "info")
-                }
-                NsSvrl.failedAssert -> {
-                    detectionFromSvrl(report, pattern, rule, node, "error")
-                }
-                else -> Unit
-            }
-        }
-    }
-
-    private fun detectionFromSvrl(report: XvrlReport, pattern: XdmNode, rule: XdmNode, node: XdmNode, defaultSeverity: String) {
-        val severity = node.getAttributeValue(_role) ?: defaultSeverity
-        val detection = report.detection(severity)
-
-        if (node.getAttributeValue(_location) != null) {
-            val location = detection.location()
-            location.xpath = node.getAttributeValue(_location)
-        }
-
-        if (rule.getAttributeValue(_context) != null) {
-            val context = detection.context()
-            val location = context.location()
-            location.xpath = rule.getAttributeValue(_context)
-        }
-
-        val lang = node.getAttributeValue(NsXml.lang)
-        for (child in node.axisIterator(Axis.CHILD)) {
-            if (child.nodeName == NsSvrl.text || child.nodeName == NsSvrl.diagnosticReference) {
-                messageFromSvrl(detection, child, lang)
-            }
-        }
-    }
-
-    private fun messageFromSvrl(detection: XvrlDetection, node: XdmNode, defaultLanguage: String?) {
-        val lang = node.getAttributeValue(NsXml.lang) ?: defaultLanguage
-        val atts = mutableMapOf<QName,String>()
-        lang?.let { atts[NsXml.lang] = it }
-        val message = detection.message(atts)
-        for (child in node.axisIterator(Axis.CHILD)) {
-            when (child.nodeKind) {
-                XdmNodeKind.TEXT -> {
-                    val text = child.stringValue.trim()
-                    if (text.isNotEmpty()) {
-                        message.message(text)
-                    }
-                }
-                XdmNodeKind.ELEMENT -> {
-                    val element = message.message(child.nodeName)
-                    recurse(element, child)
-                }
-                else -> Unit
-            }
-        }
-    }
-
-    private fun recurse(message: XvrlMessageElement, node: XdmNode) {
-        for (child in node.axisIterator(Axis.CHILD)) {
-            when (child.nodeKind) {
-                XdmNodeKind.TEXT -> {
-                    val text = child.stringValue.trim()
-                    if (text.isNotEmpty()) {
-                        message.addContent(XvrlText(stepConfig, text))
-                    }
-                }
-                XdmNodeKind.ELEMENT -> {
-                    val element = message.message(child.nodeName)
-                    recurse(element, child)
-                }
-                else -> Unit
-            }
-        }
     }
 }
