+++ xmlcalabash/src/test/kotlin/com/xmlcalabash/test/XvrlReportTest.kt
@@ -96,7 +96,7 @@ class XvrlReportTest {
         val schema = load("schema-001.xml")
         val doc = load("doc-001.xml")
         val schReport = SchematronImpl(stepConfig).report(doc, schema)
-        val reports = XvrlReports.fromSvrl(stepConfig, schReport)
+        val reports = XvrlReport.fromSvrl(stepConfig, schReport)
         val node = reports.asXml()
         Assertions.assertNotNull(node)
     }
